requirements.txt
pandas==2.0.0
dash==2.9.3
dash-bootstrap-components==1.4.1
pymongo==4.3.3
python-dotenv==1.0.0